build 33 - made hudscale floating point.

+++ trunk/paintball2/ref_gl/gl_local.h
@@ -306,9 +306,9 @@ char	*va(char *format, ...);
 void COM_StripExtension (const char *in, char *out, int out_size); // jitsecurity
 
 void	Draw_GetPicSize (int *w, int *h, char *name);
-void	Draw_Pic (int x, int y, char *name);
-void	Draw_StretchPic (int x, int y, int w, int h, char *name);
-void	Draw_Char (int x, int y, int c);
+void	Draw_Pic (float x, float y, char *name);
+void	Draw_StretchPic (float x, float y, float w, float h, char *name);
+void	Draw_Char (float x, float y, int c);
 void	Draw_TileClear (int x, int y, int w, int h, char *name);
 void	Draw_Fill (int x, int y, int w, int h, int c);
 void	Draw_FadeScreen (void);
